# fix: isWellFormed