--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -1,5 +1,8 @@
 'use strict'
 
+// eslint-disable-next-line
+const ASCII_ESCAPE = /[\u0000-\u001f\u0022\u005c]/
+
 module.exports = class Serializer {
   constructor (options) {
     switch (options && options.rounding) {
@@ -89,6 +92,7 @@ module.exports = class Serializer {
 
   asString (str) {
     const len = str.length
+
     if (len === 0) {
       return '""'
     } else if (len < 42) {
@@ -116,7 +120,7 @@ module.exports = class Serializer {
         }
       }
       return (last === -1 && ('"' + str + '"')) || ('"' + result + str.slice(last) + '"')
-    } else if (len < 5000 && str.isWellFormed()) {
+    } else if (len < 5000 && str.isWellFormed() && ASCII_ESCAPE.test(str) === false) {
       // Only use the regular expression for shorter input. The overhead is otherwise too much.
       return '"' + str + '"'
     } else {