--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -95,7 +95,13 @@ module.exports = class Serializer {
 
     if (len === 0) {
       return '""'
-    } else if (len < 42) {
+    }
+
+    if (!str.isWellFormed()) {
+      return JSON.stringify(str)
+    }
+
+    if (len < 42) {
       // magically escape strings for json
       // relying on their charCodeAt
       // everything below 32 needs JSON.stringify()
@@ -114,13 +120,13 @@ module.exports = class Serializer {
           last === -1 && (last = 0)
           result += str.slice(last, i) + '\\'
           last = i
-        } else if (point < 32 || (point >= 0xD800 && point <= 0xDFFF)) {
+        } else if (point < 32) {
           // The current character is non-printable characters or a surrogate.
           return JSON.stringify(str)
         }
       }
       return (last === -1 && ('"' + str + '"')) || ('"' + result + str.slice(last) + '"')
-    } else if (len < 5000 && str.isWellFormed() && ASCII_ESCAPE.test(str) === false) {
+    } else if (len < 5000 && ASCII_ESCAPE.test(str) === false) {
       // Only use the regular expression for shorter input. The overhead is otherwise too much.
       return '"' + str + '"'
     } else {