@Uzlopak 2025-09-27
This is imho not acceptable for fairness reasons. as the test is basically from @cesco69. You should try to give him atleast the credentials for the test.

IMHO checkout his PR by doing `gh pr checkout 788`, then create a new branch from this, as you have his commit in the git commit historie, add your solution as another commit on top of his commit.