@nigrosimone 2025-09-27
@Uzlopak I found a better combination, it seems to perform well in benchmarks
master
```
short string............................................. x 23,010,004 ops/sec ±0.58% (190 runs sampled)
unsafe short string...................................... x 1,062,255,820 ops/sec ±0.37% (191 runs sampled)
short string with double quote........................... x 13,749,470 ops/sec ±0.37% (191 runs sampled)
long string without double quotes........................ x 15,661 ops/sec ±0.31% (192 runs sampled)
unsafe long string without double quotes................. x 1,062,491,836 ops/sec ±0.37% (191 runs sampled)
long string.............................................. x 16,092 ops/sec ±0.32% (192 runs sampled)
unsafe long string....................................... x 1,060,550,353 ops/sec ±0.31% (192 runs sampled)
```

PR
```
short string............................................. x 23,078,714 ops/sec ±1.87% (178 runs sampled)
unsafe short string...................................... x 119,577,216 ops/sec ±5.05% (160 runs sampled)
short string with double quote........................... x 16,375,591 ops/sec ±2.20% (179 runs sampled)
long string without double quotes........................ x 44,909 ops/sec ±1.13% (186 runs sampled)
unsafe long string without double quotes................. x 116,764,398 ops/sec ±3.85% (164 runs sampled)
long string.............................................. x 33,744 ops/sec ±5.16% (163 runs sampled)
unsafe long string....................................... x 92,154,003 ops/sec ±5.08% (153 runs sampled)
```

 ```js
 asString (str) {
    const len = str.length

    if (len === 0) {
      return '""'
    }

    if (!str.isWellFormed()) {
      return JSON.stringify(str)
    }

    if (len < 42) {
      // magically escape strings for json
      // relying on their charCodeAt
      // everything below 32 needs JSON.stringify()
      // every string that contain surrogate needs JSON.stringify()
      // 34 and 92 happens all the time, so we
      // have a fast case for them
      let result = ''
      let last = -1
      let point = 255
      for (let i = 0; i < len; i++) {
        point = str.charCodeAt(i)
        if (
          point === 0x22 || // '"'
          point === 0x5c // '\'
        ) {
          last === -1 && (last = 0)
          result += str.slice(last, i) + '\\'
          last = i
        } else if (point < 32) {
          // The current character is non-printable characters or a surrogate.
          return JSON.stringify(str)
        }
      }
      return (last === -1 && ('"' + str + '"')) || ('"' + result + str.slice(last) + '"')
    } else if (len < 5000 && ASCII_ESCAPE.test(str) === false) {
      // Only use the regular expression for shorter input. The overhead is otherwise too much.
      return '"' + str + '"'
    } else {
      return JSON.stringify(str)
    }
  }
```